streamlit_requirements.txt
# Streamlit Dashboard Requirements
streamlit==1.31.0
pandas==2.2.0
plotly==5.18.0
kafka-python==2.0.2
scikit-learn==1.4.0
